Refactored result view controller
Split into two dedicated Result Controller. One prefixed with HUC in order to keep it backward compatible and the current one refactored to get the current info we want to expose to the merchant.
