+++ Example/PrimerSDK/Base.lproj/Main.storyboard
@@ -333,7 +333,7 @@
                                 </subviews>
                             </stackView>
                             <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Postal code:" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="Zgj-Y0-gEw">
-                                <rect key="frame" x="20" y="80" width="374" height="21"/>
+                                <rect key="frame" x="20" y="80" width="374" height="20.5"/>
                                 <fontDescription key="fontDescription" type="system" pointSize="17"/>
                                 <nil key="textColor"/>
                                 <nil key="highlightedColor"/>
@@ -628,15 +628,18 @@
                         </constraints>
                     </view>
                     <connections>
+                        <outlet property="actionsStackView" destination="jxL-iG-utE" id="b8g-eh-PWd"/>
                         <outlet property="amountLabel" destination="tFg-X3-Grv" id="8Sr-YI-Xo7"/>
+                        <outlet property="amountStackView" destination="dwO-AD-daJ" id="tIM-lG-DBl"/>
                         <outlet property="requiredActionsLabel" destination="BP7-Bj-0Wd" id="UOS-gA-ofD"/>
+                        <outlet property="responseStackView" destination="RWM-73-UkS" id="Fnh-JT-nGV"/>
                         <outlet property="responseStatus" destination="NSy-AC-wir" id="fwH-L2-lFZ"/>
                         <outlet property="responseTextView" destination="zH0-YR-jKP" id="UNG-nf-vM7"/>
                     </connections>
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="Bha-S1-DUW" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
             </objects>
-            <point key="canvasLocation" x="3719" y="133"/>
+            <point key="canvasLocation" x="3687" y="151"/>
         </scene>
         <!--Navigation Controller-->
         <scene sceneID="k2f-Pv-jbT">
