+++ Example/PrimerSDK/Data+Extensions.swift
@@ -1,0 +1,19 @@
+//
+//  Data+Extensions.swift
+//  PrimerSDK_Example
+//
+//  Created by Dario Carlomagno on 27/04/22.
+//  Copyright © 2022 CocoaPods. All rights reserved.
+//
+
+import Foundation
+
+extension Data {
+    var prettyPrintedJSONString: NSString? { /// NSString gives us a nice sanitized debugDescription
+        guard let object = try? JSONSerialization.jsonObject(with: self, options: []),
+              let data = try? JSONSerialization.data(withJSONObject: object, options: [.prettyPrinted]),
+              let prettyPrintedString = NSString(data: data, encoding: String.Encoding.utf8.rawValue) else { return nil }
+        
+        return prettyPrintedString
+    }
+}
